Split `planning.rs` code into different files within the `planning` module
